1.优化项目结构 让项目跑起来

+++ src/main/java/com/lgren/poi/poi3_17/new_read/LRowCommon.java
@@ -49,7 +49,7 @@ public class LRowCommon {
                         Method method = ReflectionUtils.findMethod(rClass, "set" + StringUtils.capitalize(field.getName()), field.getType());
                         if (method != null) {
                             if (field.getType() == Integer.class) {// TODO
-                                data = (int) data;
+                                data = data;
                             }
                             method.invoke(result, data);
                         }
